refactor: added functions for utitlity and import from index for functions
